move new-rpc to runtime clean code, remove eventChannel
