+++ packages/scene-system/sdk/runtime/SetupFpsThrottling.ts
