+++ packages/shared/apis/host/EngineAPI.ts
@@ -38,11 +38,6 @@ export function registerEngineAPIServiceServerImplementation(port: RpcServerPort
       ctx.EngineAPI.subscribedEvents[req.eventId] = false
       return {}
     },
-    async *streamEvents(req, ctx) {
-      for await (const message of ctx.eventChannel) {
-        yield { eventId: message.type, eventData: JSON.stringify(message.data) }
-      }
-    },
     async pullEvents(req, ctx) {
       const events = ctx.events.map((e) => ({ eventId: e.type, eventData: JSON.stringify(e.data) }))
       ctx.events = []
